+++ src/ac/index.js
@@ -1,51 +1,4 @@
-import {
-  ADD_ITEM,
-  ADD_REVIEW,
-  DECREMENT,
-  ERROR,
-  INCREMENT,
-  LOAD_ALL_RESTAURANTS,
-  REMOVE_ITEM,
-  SET_MIN_RATING,
-  START,
-  SUCCESS
-} from "../constants";
-
-export const increment = () => ({
-  type: INCREMENT
-});
-
-export const decrement = () => ({
-  type: DECREMENT
-});
-
-export const addItem = id => ({
-  type: ADD_ITEM,
-  payload: { id }
-});
-
-export const removeItem = id => ({
-  type: REMOVE_ITEM,
-  payload: { id }
-});
-
-export const setMinRating = minRating => ({
-  type: SET_MIN_RATING,
-  payload: { minRating }
-});
-
-export const addReview = (review, restaurantId) => ({
-  type: ADD_REVIEW,
-  payload: { review, restaurantId },
-  generateId: true
-});
-
-/*
-export const loadAllRestaurants = () => ({
-  type: LOAD_ALL_RESTAURANTS,
-  callAPI: "/api/restaurants"
-});
-*/
+import { ERROR, LOAD_ALL_RESTAURANTS, START, SUCCESS } from "../constants";
 
 export const loadAllRestaurants = () => async dispatch => {
   try {
