+++ src/constants/index.js
@@ -1,13 +1,3 @@
-export const INCREMENT = "INCREMENT";
-export const DECREMENT = "DECREMENT";
-
-export const ADD_ITEM = "ADD_ITEM";
-export const REMOVE_ITEM = "REMOVE_ITEM";
-
-export const SET_MIN_RATING = "SET_MIN_RATING";
-
-export const ADD_REVIEW = "ADD_REVIEW";
-
 export const LOAD_ALL_RESTAURANTS = "LOAD_ALL_RESTAURANTS";
 
 export const START = "_START";
