+++ src/reducer/counter/actions.js
@@ -1,0 +1,9 @@
+import { DECREMENT, INCREMENT } from "./constants";
+
+export const increment = () => ({
+  type: INCREMENT
+});
+
+export const decrement = () => ({
+  type: DECREMENT
+});
