+++ src/reducer/filters/actions.js
@@ -1,0 +1,6 @@
+import { SET_MIN_RATING } from "./constants";
+
+export const setMinRating = minRating => ({
+  type: SET_MIN_RATING,
+  payload: { minRating }
+});
