+++ src/reducer/order/actions.js
@@ -1,0 +1,11 @@
+import { ADD_ITEM, REMOVE_ITEM } from "./constants";
+
+export const addItem = id => ({
+  type: ADD_ITEM,
+  payload: { id }
+});
+
+export const removeItem = id => ({
+  type: REMOVE_ITEM,
+  payload: { id }
+});
