+++ src/reducer/reviews/actions.js
@@ -1,0 +1,7 @@
+import { ADD_REVIEW } from "./constants";
+
+export const addReview = (review, restaurantId) => ({
+  type: ADD_REVIEW,
+  payload: { review, restaurantId },
+  generateId: true
+});
