+++ src/reducer/reviews/constants.js
@@ -1,0 +1,1 @@
+export const ADD_REVIEW = "ADD_REVIEW";
