+++ src/reducer/reviews/reducer.js
@@ -1,6 +1,6 @@
-import { normalizedReviews } from "../fixtures";
-import { arrToMap } from "../utils";
-import { ADD_REVIEW } from "../constants";
+import { normalizedReviews } from "../../fixtures";
+import { arrToMap } from "../../utils";
+import { ADD_REVIEW } from "./constants";
 
 const defaultReviews = arrToMap(normalizedReviews);
